+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changes
 =======
 
+dev (master)
+------------
+
+* Change ``HTTPSConnection`` to load system CA certificates
+  when ``ca_certs``, ``ca_cert_dir``, and ``ssl_context`` are
+  unspecified. (Pull #)
+
+
 1.25.2 (2019-04-28)
 -------------------